// src/libs/data-access/klass/models/CorrespondenceTableSummaryResource.ts
export interface CorrespondenceTableSummaryResource {
    /**
     * 
     * @type {string}
     * @memberof CorrespondenceTableSummaryResource
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof CorrespondenceTableSummaryResource
     */
    id?: number;
    /**
     * 
     * @type {ContactPersonResource}
     * @memberof CorrespondenceTableSummaryResource
     */
    contactPerson?: ContactPersonResource;
    /**
     * 
     * @type {string}
     * @memberof CorrespondenceTableSummaryResource
     */
    owningSection?: string;
    /**
     * 
     * @type {string}
     * @memberof CorrespondenceTableSummaryResource
     */
    source?: string;
    /**
     * 
     * @type {number}
     * @memberof CorrespondenceTableSummaryResource
     */
    sourceId?: number;
    /**
     * 
     * @type {string}
     * @memberof CorrespondenceTableSummaryResource
     */
    target?: string;
    /**
     * 
     * @type {number}
     * @memberof CorrespondenceTableSummaryResource
     */
    targetId?: number;
    /**
     * 
     * @type {boolean}
     * @memberof CorrespondenceTableSummaryResource
     */
    changeTable?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof CorrespondenceTableSummaryResource
     */
    lastModified?: Date;
    /**
     * 
     * @type {Array<string>}
     * @memberof CorrespondenceTableSummaryResource
     */
    published?: Array<string>;
    /**
     * 
     * @type {LevelResource}
     * @memberof CorrespondenceTableSummaryResource
     */
    sourceLevel?: LevelResource;
    /**
     * 
     * @type {LevelResource}
     * @memberof CorrespondenceTableSummaryResource
     */
    targetLevel?: LevelResource;
    /**
     * 
     * @type {{ [key: string]: Link; }}
     * @memberof CorrespondenceTableSummaryResource
     */
    links?: { [key: string]: Link; };
    /**
     * 
     * @type {ClassificationVariantSummaryResourceEmbedded}
     * @memberof CorrespondenceTableSummaryResource
     */
    embedded?: ClassificationVariantSummaryResourceEmbedded;
    /**
     * 
     * @type {PageMetadata}
     * @memberof CorrespondenceTableSummaryResource
     */
    page?: PageMetadata;
}